CodeDoctor_khtml-c4ffcccf175e75f70f4253e58ac5ea1024b21878/requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=3.4.0
